requirements.txt
flask==2.3.3
werkzeug==2.3.7
openai>=1.3.0
python-dotenv==0.19.0
